backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
google-api-python-client
google-auth
